# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "PSP System App",
        "contact": {
            "name": "Mahmoud Mohamed",
            "url": "https://www.linkedin.com/in/mahmoud-mohamed-955932b5/",
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/{guard}/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "patient"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/{guard}/reset-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "reset password",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "rep1@gmail.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/{guard}/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout currently logged in user",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/{guard}/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "refreshes expired token",
                "parameters": [
                    {
                        "name": "guard",
                        "in": "path",
                        "description": "userApi , companyApi , pharmacyAdminApi , pharmacyUsersApi ",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "userApi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "403": {
                        "description": "Forbidden The client did not have permission to access the requested resource."
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}